[Media Router] Add a policy to enable/disable Media Router.
- Add policy "EnableMediaRouter". The default (if not set) will be to
enable Media Router.
-- EnableMediaRouter will be a user level policy that affects Chrome,
Chrome OS, and Android.
-- Mandatory policy.
- Add logic in media_router::MediaRouterEnabled() to check if the
corresponding preference is set from a mandatory policy. If so, return
that value. Otherwise, check the existing FeatureSwitch (or return true
on Android). If not set, or if not set from mandatory policy, then the
preference is ignored.
- Reordered the logic in the function to share common logic between
desktop and Android.

BUG=558453

Review-Url: https://codereview.chromium.org/2035903002
Cr-Original-Commit-Position: refs/heads/master@{#398377}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c2b6c183b1817720f946b3c6378e3479d9083722