tools/android/loading Move task creation to background and add dashboard
This CL moves the tasks creation to the background queue,
to avoid AppEngine timeouts.

It also adds a way to list the current jobs and track them.

The list of jobs is persisted in the datastore.

There is more work to be done in follow up CLs:
- improving the UI for the job dashboard
- possibly replacing the per-job polling by a global
  polling that would poll all active jobs at once, in order
  to simplify the code.

Review-Url: https://codereview.chromium.org/2041193008
Cr-Original-Commit-Position: refs/heads/master@{#399175}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4953d75f1ebef626ca9dfcc20630193dedcfb8ea