[Origin Trials] Improve developer tools ergonomics
This CL fixes two pain points with the generate_token.py tool:
1. Prints out the token details with the token itself (since we moved to v2
tokens, it's impossible to see at a glance what origin and feature a token is
for.)
2. Rejects hostnames without a '.' character, with a special exception for
localhost. This is becuase hostname and feature name are positional arguments,
and it's easy to swap them accidentally and produce broken tokens.

Review-Url: https://codereview.chromium.org/2064393003
Cr-Original-Commit-Position: refs/heads/master@{#400108}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08b9e8da6788c2e8577571eb61a5b70a4c5e5222