Reland: [PCv2] Add tracing category needed for computing ttfmp
The time-to-first-meaningful-paint computation requires extra tracing categories to be enabled.
This CL enables recording of the category in PCv2.

https://github.com/catapult-project/catapult/blob/master/tracing/tracing/metrics/system_health/first_paint_metric.html

This is a reland of https://codereview.chromium.org/2049593002/ with
disabling the benchmark on reference build. In the reference build, the
disabled-by-default-blink.debug.layout category brows up trace size and
causes d8 crash.

BUG=615605,619254

Review-Url: https://codereview.chromium.org/2070773002
Cr-Original-Commit-Position: refs/heads/master@{#400426}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e598d9f6f9d6bf4cd90c46503cb0c207791a02db