[Custom Tabs] Implement "Read It Later" Behind a Flag
This CL implements "read it later" feature and puts the feature behind a
flag. The trial name is "ReadItLater". When this mode is turned on,
reparented tab will stay visible for 600MS and then the activity is
pushed to back of the recent stack.

BUG=620535

Review-Url: https://codereview.chromium.org/2076463002
Cr-Original-Commit-Position: refs/heads/master@{#400524}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea63767b61ae1a8935b420685bbee58467b68b5b