This fix uses Optional TimeDeltas since they may be non existent instead of using zero as a non-existent value. It uses base::Optional class (https://codereview.chromium.org/1245163002/) BUG=596367
Apart from the initial fix for this change which involves TimeDelta values in
PageLoadExtraInfo to be converted to use base::Optional, this CL has the following
3 more changes:
1. Converted the foreground event condition from event_time < background_time to
event_time <= background_time. This made all tests that have an event occuring
before calling WasHidden (which fills the background_time as Now()) to be
deterministic in assuming that the event will be logged in a foreground histogram
irrespective of the clock resolution. This is also a better logical choice since
the event has not been subjected to any background-specific functionality if it
occurred at the same time as backgrounding. This fixed the first failure reported
on this CL :CorePageLoadMetricsObserverTest.OnlyBackgroundLaterEvent.

The tests that have event occurring after background will be non
deterministic as they cannot be guaranteed to have event time strictly greater
than background time.
So I have added an interface in the test
framework that lets the test retrieve the value of PageLoadExtraInfo for a
committed load. I have not added the same for provisional loads/aborted loads but
could be added in future if need be.

2. For the failure resulting from inter process time tick skew, conditional
clamping is introduced only if the clock is not high resolution. The testing
strategy for that is that the existing unit tests should all pass on a low
resolution clock system.  Failures from this should result in a DCHECK getting
hit.
(I have not added specific tests for testing this because
it will be difficult to test all possible code paths in specific tests and mocking
inter process time tick skew would require mocking isHighResolution at the least.
The complexity is not worth it since all code paths will not get covered in a few
tests anyways.)

Review-Url: https://codereview.chromium.org/1837233002
Cr-Original-Commit-Position: refs/heads/master@{#401022}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 271f0336503725749c86df213f6710c6423f9b5f