media: Add RAPPOR reporting for video codecs using SRC and MSE
This CL reports URLs that uses a specific video codec in FFmpegDemuxer
(for SRC playback) and ChunkDemuxer (for MSE playback). All video
codecs will be reported, but we are only interested in a few of them in
rappor.xml (h264, vp8 and vp9).

BUG=499949
TEST=Manually tested and saw the metric reported.

Review-Url: https://codereview.chromium.org/2087973002
Cr-Original-Commit-Position: refs/heads/master@{#401416}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4abfbdc1216dbaf686bafc07f1de974b2cb05f14