Report precache metrics even if native library is not loaded
Metrics recorded in precache java code when the native library is not
loaded are lost. The metrics are recorded in shared preferences and
recorded to UMA when the library successfully loads in future.

Recommitting CL https://codereview.chromium.org/1988513003/ which got
reverted due to some instrumentation test failures. The fix is in
PrecacheUMA.java:132

BUG=605734

Review-Url: https://codereview.chromium.org/2098063003
Cr-Original-Commit-Position: refs/heads/master@{#402488}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 52ae19d6ec856c34195f1e8f0de6e2b6d1675c66