Supervised users initiated extensions installs
Supervised users can install apps and extension.  The extension will be
installed in a disabled state and an approval request will be sent to
their custodian.

If the custodian approves the request, the extension will be enabled.
If the custodian denies the request, the extension will be uninstalled.

BUG=619526

Review-Url: https://codereview.chromium.org/2004043002
Cr-Original-Commit-Position: refs/heads/master@{#402570}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e960964b7ea2caff5289d4a2f40488af7613389d