Add time to parse start broken out by load type.
This will allow us to better understand the potential
for improvement if we serve all back/forward nav main
resources from cache.

We may wish to follow this change with a change that
logs this metric, but just for pages where the main
resource is no-store.

Review-Url: https://codereview.chromium.org/2103363003
Cr-Original-Commit-Position: refs/heads/master@{#402928}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de8577f88886e6dc97e9fcc88e372547dbdf12a8