[TTS] Basic Tap Suppression functionality.
Adds a feature to suppress the Contextual Search UX when a Tap
occurs and various signals indicate that it's unlikely the user
will open the panel.

For this CL, the only two signals are:
1) Whether the user has Tapped N times without opening the panel.
2) Whether the previous Tap was in about the same location recently.

Signal #1 suppresses, signal #2 overrides the suppression by triggering.

Also minor cleanup to TapFarFromPreviousSuppression and
ContextualSearchTapState.

BUG=609918

Review-Url: https://codereview.chromium.org/2096203002
Cr-Original-Commit-Position: refs/heads/master@{#403239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: efdef4bf3549ba781852f35b4b642a21d9878b3b