[sql] Database recovery system for Shortcuts.
sql::Recovery provides a conservative recovery system which requires
some overhead in the client code.  Such code exists for the "Top Sites"
and Favicons databases, and has shown that the easy-to-handle cases
dominate.  Implement a helper function to recover an entire database,
and apply it to the Shortcuts database.

BUG=597785

Review-Url: https://codereview.chromium.org/1832173002
Cr-Original-Commit-Position: refs/heads/master@{#403576}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a402e75a80c13fab013100447b31a00ff6b3a0f5