CWS private API (webstorePrivate.isPendingCustodianApproval)
Adding a new CWS private API (webstorePrivate.isPendingCustodianApproval)
to check if an extension is pending custodian approval.  It is used in
the wesbtore to show the extension as "Pending" when it has been installed
or updated and the custodian hasn't approved it yet.
Custodian approvals are required for extensions that are installed by
supervised users and Unicorn users, or for updates that require additional
permissions.

BUG=609849
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:closure_compilation

Review-Url: https://codereview.chromium.org/1948133004
Cr-Original-Commit-Position: refs/heads/master@{#403674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7715a2ae33d1bcc0515b47254c305e47d0b6a18