Add metrics for app banner preventDefault() and prompt().
The beforeinstallprompt event signals an imminent app install banner for
a page. This CL adds UMA stats for preventDefault() being called on the
event (stopping the banner from appearing), and for prompt() being
called on the event (making the banner appear subsequent to
a preventDefault()).

BUG=625716

Review-Url: https://codereview.chromium.org/2122783002
Cr-Original-Commit-Position: refs/heads/master@{#404039}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f6382c45e105322038b6fa1885cd1f479ef5ff8