Mark build arguments implicitly used in GN.
This regressed in https://codereview.chromium.org/2092623002 which caused a revert of the GN roll. The particular issue is that a build argument in a BUILD.gn file was used in only one of the toolchains, so the other toolchains were giving unused variable errors on it.

Build variables kept at their default values already are marked implicitly used, and anything in a .gni file is the same. The remaining cases are pretty confusing so it seems better to not warn in this case instead of requiring the author to wrap the build variable in a condition based on the usage.

TBR=dpranke

Review-Url: https://codereview.chromium.org/2151633002
Cr-Original-Commit-Position: refs/heads/master@{#405291}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8b05f575431f4854163470695cf1a6f6495263ce