[Extensions] Add extension feature generation code
Extension features files are json files that contain information about
which APIs or extension features can be used in which contexts.
Currently, these files are parsed at runtime during both browser
startup and renderer process startup, causing a significant delay.

This patch implements the bulk of the functionality to move towards
generating feature files rather than parsing them from the json, which
should hopefully result in a significant speed improvement.

This introduces the feature_compiler.py, which generates the .h and .cc
files for a given set of features.json files, as well as the build rules
to invoke it automatically and tests for it. This does not add generated
files to the production build of chrome yet.

Sample api_feature_provider.h and .cc:
https://gist.github.com/anonymous/9bb094a90810b6050556761c87c204ac

BUG=280286

Review-Url: https://codereview.chromium.org/2151583003
Cr-Original-Commit-Position: refs/heads/master@{#406575}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88155eb86d2b8b8493eb80d09feb639d34dfdcc4