Implement multi-phase build support in MB.
The PGO builders need to run multiple compiles in a single build,
each with different arguments. This was not previously possible
with MB, but this CL adds support for specifying a list of configs
per builder, and adds a --phase command line arg to indicate which
phase of the build should be generated.

It is an error to specify --phase on a builder that isn't
expected to have multiple phases. Also phases are 1-based (phase 1,
phase 2, etc.).

Also, this CL cleans up some TODOs that I no longer plan to fix.

R=sebmarchand@chromium.org
BUG=595947

Review-Url: https://codereview.chromium.org/2160353003
Cr-Original-Commit-Position: refs/heads/master@{#406955}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9380a1bd84d8be46835457a9d2a9c004d02e344