Measure how often SCTs can't be checked because they're too new
Signed Certificate Timestamps (SCTs) designate Merkle tree leaves that can
be checked for inclusion with a CT Log's Signed Tree Head (STH).
However, inclusion should only be checked against an STH that is newer
than the SCTs timestamp by at least 24 hours - this is because Logs'
have a Maximum Merge Delay of 24 hours, which is the time they have
to produce a new STH that incorporates a given SCT.

STHs are delivered periodically out of band. If there isn't a new enough
STH, then SCTs will need to be marked as pending inclusion check,
waiting for a new STH to be delivered.

To determine how frequently an STH should be delivered (currently,
daily) and how big the SCT queue list should be (that is, how frequently
clients encounter a brand new SCT that is newer than any STH),
measure how often an SCT can't be checked for inclusion immediately
after it's been seen.

BUG=613495

Review-Url: https://codereview.chromium.org/2153123002
Cr-Original-Commit-Position: refs/heads/master@{#407471}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b03a8a77ead185c85a6c55303ac20176eaa20a56