Notify page load metrics when the app enters the background.
This change adds a PageLoadMetricsProvider, which implements the new
OnAppEnterBackground hook, and informs all MetricsWebContentsObservers
and in turn all PageLoadTrackers when the app enters the background.

On Android (and iOS), when an app enters the background, it may be
killed without further notifications. Thus, having this callback
will allow us to persist metrics before the app may be killed.
Our page load metrics subsystem isn't used on iOS, so we only target
Android in this change.

Unfortunately, we can't use the existing WebContents hidden callback,
as this gets invoked after the UMA subsystem has already flushed
metrics to disk. Thus, we need to add this explicit callback which
allows us to persist metrics as part of the UMA subsystem handling
the backgrounding callback.

For now, we log a debugging histogram in PageLoadTracker to help
understand how often trackers enter the background, and how often
those trackers are destroyed without running their destructors. In a
subsequent change, we'll add support for informing PLMObservers that
the app entered the background. This will allow any PLMObsevers that
implement OnComplete or OnFailedProvisionalLoad to also get called back
in cases where the tracker may be destroyed before its destructor runs,
due to the application being killed.

BUG=608360

Review-Url: https://codereview.chromium.org/2189543002
Cr-Original-Commit-Position: refs/heads/master@{#408744}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9eca220417aa1050fbe8b431fee890f0b5b42240