Strip comments and whitespace from javascript resources
On Android strip comments and whitespace from javascript resources.
Add an option to grit build that provides a command that strips
comments and whitespace, and modify grit so that this is applied
to all javascript resource and scripts within HTML resources.
Add a python script that wraps the closure compiler to implement
this command.
Pull it all together with GN.

BUG=619091

Review-Url: https://codereview.chromium.org/2179033002
Cr-Original-Commit-Position: refs/heads/master@{#409498}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dfa12f1b8845bc912728471d19d9915119456d7