Rice decode removals and additions and use them to update the store.
Also:
1. Add histograms for decoding additions/removals.
2. Update the histogram for ApplyUpdateResult to include the new failure values.

Please note that this CL does not enable RICE encoding yet since we still
telling the server that we can only process RAW updates. I'll make that tiny
change as a separate CL so that it is easier to rollback, if needed.

BUG=543161,624567

Review-Url: https://codereview.chromium.org/2190233002
Cr-Original-Commit-Position: refs/heads/master@{#409844}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 605ab1c3447ebe4749d80c20a7d38c01a59c0887