Reduce warnings about non-literal histogram names
find_unmapped_histograms.py emits warnings when it encounters a
histogram name that is not a literal string. This change increases the
set of names that are considered safe.

1. Don't warn for constant names containing multiple namespaces,
eg. user_prefs::tracked::kTrackedPrefHistogramReset
2. Don't warn for empty names, eg. UMA_HISTOGRAM_ENUMERATION(). This must
be in a comment or otherwise non-executable, or it wouldn't compile.
3. Don't warn for macro concatenations. These are always constant and
therefore safe.

R=isherman@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2224273002
Cr-Original-Commit-Position: refs/heads/master@{#411546}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8b0f987a888348117208cb697742c5fed0d859fd