Adding SwiftShader to the DEPS file
The SwiftShader library has recently been made open source
and can now be pulled with the Chromium checkout. The size
of the SwiftShader library is roughly 162 Mb, currently.
Note that this will eventually allow us to remove MESA,
which takes approximately 66 Mb, so the net gain should be
under 100Mb in total.

Pulling SwiftShader along with Chromium will allow bots to
build SwiftShader, perform tests with the SwiftShader
library and have quick turnaround time when issues are
found.

This will not change the state of SwiftShader as a
component in Chrome and it will not be shipped with Chrome,
only downloaded as a component on demand.

Also note that this cl does not enable building
SwiftShader, it only allows the SwiftShader library to be
pulled, to make sure that everyone agrees with this first
step.

This is a re-land of https://codereview.chromium.org/2168143003/
after having fixed ozone build breakage (Issue 635702) with
https://codereview.chromium.org/2231913002/

BUG=630728
TBR=jam@chromium.org,phajdan.jr@chromium.org

Change-Id: Iac26a520fffa38410e81f49c666c17e7ceecfbc2
Review-Url: https://codereview.chromium.org/2238063003
Cr-Original-Commit-Position: refs/heads/master@{#411793}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e3976c2e7d2ce17180ebf6d2a576f14c188c843d