Measure <body onselectionchange> content attribute
This is the only event handler attribute on body that forwards to the
document instead of the window. WebKit also has it as an IDL attribute,
while Edge and Gecko don't have onselectionchange on body at all.

This is in order to inform the spec:
https://github.com/w3c/selection-api/issues/54#issuecomment-239574361

R=tkent@chromium.org,isherman@chromium.org

Review-Url: https://codereview.chromium.org/2249613002
Cr-Original-Commit-Position: refs/heads/master@{#412824}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d50359eabc1cf54bfc52be4cf4e19b8b4dceb9f