predictors: Track whether resources have validators, and require validation.
This adds to the database whether resources have validators, and are
never fresh because they require revalidation each time. This will be
used later to adjust the priority.

Also adds tests for SummarizeResponse, and unifies (and fixes) the
content type handling in it.

BUG=631966

Review-Url: https://codereview.chromium.org/2260573002
Cr-Original-Commit-Position: refs/heads/master@{#413202}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c65a3bf951519f90b6be30a1bded64a7001e59d0