Add support for the ExperimentalSwReporterEngine field trial.
If the field trial is enabled, read a "tag" value from the variation parameters
and pass it on to the component updater. After the experimental component is
installed, read command-line arguments and a "suffix" string from its manifest;
pass the command-line arguments to the SwReporter when it's launched, and
append the suffix to all histograms it logs to distinguish them from reports
from the default reporter.

BUG=631480

Review-Url: https://codereview.chromium.org/2278013002
Cr-Original-Commit-Position: refs/heads/master@{#414606}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4b1fb47a28f8c3558ac10cb04be327fa7e228ae