Add a metric measuring how many times an infobar hid other infobars on Android.
This CL adds a new UMA metric to record the identifier of an infobar
which was visible to the user and hid another infobar behind it. This CL
also updates the visible/hidden check to simply query whether or not
another infobar exists. This is to correct an overcounting issue that
recorded infobars which appeared after another infobar was dismissed
as being hidden.

BUG=589163

Review-Url: https://codereview.chromium.org/2321923002
Cr-Original-Commit-Position: refs/heads/master@{#417496}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 563b9b34f5dfba7efc69e3c261989a6f323c8cd2