[Offline Pages] Adds duplicate detection metrics.
We want to know whether the UI for showing page downloads as starting
is sufficient to let the user know that something is happening.  The
hypothesis is that if users get too many duplicate pages then they aren't
noticing that the pages are getting saved on the first button click.

BUG=643748

Review-Url: https://codereview.chromium.org/2314273003
Cr-Original-Commit-Position: refs/heads/master@{#418331}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e3105ed9096bf45f4ca2d79b4a682adc1d1d9ae